requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
netifaces==0.10.9
progressbar2==3.51.4
psutil==5.7.2
PyAccessPoint==0.2.5
python-utils==2.4.0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
wireless==0.3.3
